Adding create wallet button to the toolbar
