+++ app/frontend/actions.js
@@ -154,6 +154,7 @@ module.exports = ({setState, getState}) => {
       mnemonic: generateMnemonic(),
       mnemonicValidationError: undefined,
       showGenerateMnemonicDialog: true,
+      authMethod: 'mnemonic',
     })
   }
