+++ app/frontend/components/pages/login/mnemonicAuth.js
@@ -2,15 +2,7 @@ const {h, Component} = require('preact')
 const {getTranslation} = require('../../../translations')
 const connect = require('unistore/preact').connect
 const actions = require('../../../actions')
-
-const GenerateMnemonicDialog = require('./generateMnemonicDialog')
-
-const isLeftClick = (e, action) => {
-  if (e.button === 0) {
-    e.preventDefault()
-    action()
-  }
-}
+const isLeftClick = require('../../../helpers/isLeftClick')
 
 class LoadByMenmonicSectionClass extends Component {
   componentDidMount() {
@@ -31,10 +23,8 @@ class LoadByMenmonicSectionClass extends Component {
     updateMnemonic,
     checkForMnemonicValidationError,
     loadWallet,
-    showGenerateMnemonicDialog,
     loadDemoWallet,
     showMnemonicValidationError,
-    openGenerateMnemonicDialog,
   }) {
     return h(
       'div',
@@ -92,19 +82,6 @@ class LoadByMenmonicSectionClass extends Component {
         )
       ),
       h(
-        'a',
-        {
-          class: 'intro-link',
-          /*
-          * onMouseDown to prevent onBlur before handling the click event
-          * https://stackoverflow.com/questions/17769005/onclick-and-onblur-ordering-issue
-          */
-          onMouseDown: (e) => isLeftClick(e, openGenerateMnemonicDialog),
-        },
-        '…or generate a new one'
-      ),
-      showGenerateMnemonicDialog && h(GenerateMnemonicDialog),
-      h(
         'div',
         {class: 'centered-row'},
         h(
@@ -130,7 +107,6 @@ module.exports = connect(
     displayAboutOverlay: state.displayAboutOverlay,
     showDemoWalletWarningDialog: state.showDemoWalletWarningDialog,
     mnemonicValidationError: state.mnemonicValidationError,
-    showGenerateMnemonicDialog: state.showGenerateMnemonicDialog,
     showMnemonicValidationError: state.showMnemonicValidationError,
   }),
   actions
