+++ app/frontend/helpers/isLeftClick.js
@@ -1,0 +1,8 @@
+const isLeftClick = (e, action) => {
+  if (e.button === 0) {
+    e.preventDefault()
+    action()
+  }
+}
+
+module.exports = isLeftClick
